Calibrate 100% functional, setSpeedsRPS functional.  all commits by this account because reasons

+++ libraries/Code/MyEncoders.h
@@ -19,7 +19,7 @@ void getCounts(unsigned long counts[]);
 
 //this function should return the instantaneous speeds of the wheels
 //meassured in revolutions per second.
-void getSpeeds(float speeds[]);
+void getSpeeds(double speeds[]);
 
 //this function should include whatever code necessary to initialize this module
